# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Store API",
        "description": "API для работы с книгами в интернет-магазине",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http:///localhost:8000/api",
            "description": "Production server"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Получить список книг в продаже",
                "description": "Возвращает отфильтрованный и отсортированный список книг",
                "operationId": "0f3615f70ddcc86b61526e34a8444165",
                "parameters": [
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки (доступные значения: id, title - название, price - цена, quantity - количество, created_at - дата создания, sells_count - количество продаж)",
                        "schema": {
                            "type": "string",
                            "default": "id",
                            "enum": [
                                "id",
                                "title",
                                "price",
                                "quantity",
                                "created_at",
                                "sells_count"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "min_quantity",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "with_author_avatar",
                        "in": "query",
                        "description": "Показать только книги авторов с аватарками",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooksListResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Получить детальную информацию о книге",
                "operationId": "f220165fa9917d68dcc608e5197af008",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookDetailResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Книга не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/{id}/purchase": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Купить книгу",
                "operationId": "98e37fdeb2240c4b296635888af5c296",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "default": 1,
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Книга успешно куплена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при покупке",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Книга не найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/books/top": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Получить список книг авторов с высоким рейтингом или с высокими продажами за сегодня",
                "description": "Возвращает список книг у авторов которых рейтинг выше 75 либо количество продаж за текущий день больше 3",
                "operationId": "77bd082372aca2d099ef0a620a29addb",
                "parameters": [
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки (доступные значения: id, title - название, price - цена, quantity - количество, created_at - дата создания, sells_count - количество продаж)",
                        "schema": {
                            "type": "string",
                            "default": "id",
                            "enum": [
                                "id",
                                "title",
                                "price",
                                "quantity",
                                "created_at",
                                "sells_count"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_price",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "max_price",
                        "in": "query",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "min_quantity",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_author_rank",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 75
                        }
                    },
                    {
                        "name": "min_today_sales",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный запрос",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BooksListResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Название книги"
                    },
                    "description": {
                        "type": "string",
                        "example": "Описание книги"
                    },
                    "cover_url": {
                        "type": "string",
                        "example": "https://example.com/cover.jpg"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 10
                    },
                    "authors": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Author"
                        }
                    },
                    "sells_count": {
                        "description": "Количество продаж книги",
                        "type": "integer",
                        "example": 15
                    },
                    "in_stock": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Author": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "first_name": {
                        "type": "string",
                        "example": "Иван"
                    },
                    "last_name": {
                        "type": "string",
                        "example": "Иванов"
                    },
                    "full_name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "rank": {
                        "type": "integer",
                        "example": 5
                    },
                    "avatar_url": {
                        "type": "string",
                        "example": "https://example.com/avatar.jpg"
                    },
                    "books_count": {
                        "type": "integer",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Sale": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "book": {
                        "$ref": "#/components/schemas/Book"
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 19.99
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "BooksListResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Book"
                        }
                    },
                    "meta": {
                        "properties": {
                            "filters": {
                                "type": "object"
                            },
                            "sort": {
                                "properties": {
                                    "field": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "direction": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "total_count": {
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "BookDetailResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "$ref": "#/components/schemas/Book"
                    }
                },
                "type": "object"
            },
            "PurchaseResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "sale_id": {
                                "type": "integer"
                            },
                            "book": {
                                "$ref": "#/components/schemas/Book"
                            },
                            "quantity": {
                                "type": "integer"
                            },
                            "total_price": {
                                "type": "number",
                                "format": "float"
                            },
                            "sale": {
                                "$ref": "#/components/schemas/Sale"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "properties": {
                            "available": {
                                "type": "integer"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Books",
            "description": "Операции с книгами"
        }
    ]
}